Expanded event icon display

+++ code/src/net/filmos/az/colors/ColorGradient.java
@@ -11,20 +11,20 @@ public class ColorGradient {
         this.right = right;
     }
 
-    public Color getColor(float ratio) {
+    public Color getColor(double ratio) {
         return interpolateColor(left, right, ratio);
     }
 
-    public static Color interpolateColor(Color left, Color right, float ratio) {
+    public static Color interpolateColor(Color left, Color right, double ratio) {
         ratio = Math.max(0, Math.min(1, ratio));
 
-        float leftR = convertToLinearChannel(left.getRed()/255f);
-        float leftG = convertToLinearChannel(left.getGreen()/255f);
-        float leftB = convertToLinearChannel(left.getBlue()/255f);
+        double leftR = convertToLinearChannel(left.getRed()/255f);
+        double leftG = convertToLinearChannel(left.getGreen()/255f);
+        double leftB = convertToLinearChannel(left.getBlue()/255f);
 
-        float rightR = convertToLinearChannel(right.getRed()/255f);
-        float rightG = convertToLinearChannel(right.getGreen()/255f);
-        float rightB = convertToLinearChannel(right.getBlue()/255f);
+        double rightR = convertToLinearChannel(right.getRed()/255f);
+        double rightG = convertToLinearChannel(right.getGreen()/255f);
+        double rightB = convertToLinearChannel(right.getBlue()/255f);
 
         return new Color(
             Math.round(convertFromLinearChannel(leftR+ratio*(rightR-leftR))*255),
@@ -33,13 +33,13 @@ public class ColorGradient {
         );
     }
 
-    static float convertToLinearChannel(float srgb) {
+    static double convertToLinearChannel(double srgb) {
         // IEC 61966-2-1:1999
-        return srgb <= 0.04045f ? srgb / 12.92f : (float) pow((srgb + 0.055f) / 1.055f, 2.4f);
+        return srgb <= 0.04045d ? srgb / 12.92d : pow((srgb + 0.055d) / 1.055d, 2.4d);
     }
-    static float convertFromLinearChannel(float linear) {
+    static double convertFromLinearChannel(double linear) {
         // IEC 61966-2-1:1999
-        return linear <= 0.0031308f ?
-                linear * 12.92f : (float) ((pow(linear, 1.0f / 2.4f) * 1.055f) - 0.055f);
+        return linear <= 0.0031308d ?
+                linear * 12.92d : ((pow(linear, 1.0d / 2.4d) * 1.055d) - 0.055d);
     }
 }
